+++ src/model/comment_report.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type CommentReport struct {
-	Id             int64     `json:"id" gorm:"default"`
-	CommentId      int64     `json:"comment_id"`
-	UserId         int64     `json:"user_id"`
+	Id             int       `json:"id" gorm:"default"`
+	CommentId      int       `json:"comment_id"`
+	UserId         int       `json:"user_id"`
 	ReportCategory string    `json:"report_category"`
 	Comment        string    `json:"comment"`
 	Date           time.Time `json:"date" gorm:"default"`
