+++ src/model/proposal_like.go
@@ -2,10 +2,10 @@ package model
 
 import "time"
 
-type ProposalLike struct{
-	Id 	       int64  	 `json:"id" gorm:"default"`
-	ProposalId int64     `json:"proposal_id"`
-	UserId     int64  	 `json:"user_id"` 
+type ProposalLike struct {
+	Id         int       `json:"id" gorm:"default"`
+	ProposalId int       `json:"proposal_id"`
+	UserId     int       `json:"user_id"`
 	Feedback   string    `json:"feedback"`
 	Date       time.Time `json:"date" gorm:"default"`
 }
