+++ src/model/proposal_report.go
@@ -2,11 +2,11 @@ package model
 
 import "time"
 
-type ProposalReport struct{
-	Id 	       	   int64  	 `json:"id" gorm:"default"`
-	ProposalId 	   int64     `json:"proposal_id"`
-	UserId     	   int64  	 `json:"user_id"` 
+type ProposalReport struct {
+	Id             int       `json:"id" gorm:"default"`
+	ProposalId     int       `json:"proposal_id"`
+	UserId         int       `json:"user_id"`
 	ReportCategory string    `json:"report_category"`
-	Comment	   	   string 	 `json:"comment"`
-	Date       	   time.Time `json:"date" gorm:"default"`
+	Comment        string    `json:"comment"`
+	Date           time.Time `json:"date" gorm:"default"`
 }
