+++ src/model/report_category.go
@@ -1,6 +1,6 @@
 package model
 
 type ReportCategory struct {
-	Id   int64  `json:"id" gorm:"default"`
+	Id   int    `json:"id" gorm:"default"`
 	Name string `json:"name"`
 }
